# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vowed API",
    "version": "0.1.0",
    "description": "Backend for the Vowed Android app. All endpoints except /v1/health, /v1/meta and /v1/auth/* need `Authorization: Bearer <token>` from /v1/auth/verify. u64 amounts are decimal strings. Transaction builders return unsigned base64 transactions: the app must decode and check them before asking the wallet to sign."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/squads": {
      "post": {
        "summary": "Create a squad",
        "tags": [
          "squads"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "owner": {
                      "type": "string"
                    },
                    "inviteCode": {
                      "type": "string"
                    },
                    "deepLink": {
                      "type": "string"
                    },
                    "memberCount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "owner",
                    "inviteCode",
                    "deepLink",
                    "memberCount"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "My squads",
        "tags": [
          "squads"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "squads": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "owner": {
                            "type": "string"
                          },
                          "inviteCode": {
                            "type": "string"
                          },
                          "deepLink": {
                            "type": "string"
                          },
                          "memberCount": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "owner",
                          "inviteCode",
                          "deepLink",
                          "memberCount"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "squads"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/squads/join": {
      "post": {
        "summary": "Join a squad with an invite code",
        "tags": [
          "squads"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^[A-Z0-9]{8}$"
                  }
                },
                "required": [
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "owner": {
                      "type": "string"
                    },
                    "inviteCode": {
                      "type": "string"
                    },
                    "deepLink": {
                      "type": "string"
                    },
                    "memberCount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "owner",
                    "inviteCode",
                    "deepLink",
                    "memberCount"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/squads/{id}": {
      "get": {
        "summary": "Squad details with members and linked challenges",
        "tags": [
          "squads"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{32}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "squad": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "owner": {
                          "type": "string"
                        },
                        "inviteCode": {
                          "type": "string"
                        },
                        "deepLink": {
                          "type": "string"
                        },
                        "memberCount": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "owner",
                        "inviteCode",
                        "deepLink",
                        "memberCount"
                      ],
                      "additionalProperties": false
                    },
                    "members": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "wallet": {
                            "type": "string"
                          },
                          "joinedAt": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "wallet",
                          "joinedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "challenges": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "pool": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "startTs": {
                            "type": "number"
                          },
                          "endTs": {
                            "type": "number"
                          },
                          "durationDays": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "pool",
                          "status",
                          "startTs",
                          "endTs",
                          "durationDays"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "squad",
                    "members",
                    "challenges"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/squads/{id}/invite": {
      "post": {
        "summary": "Get the invite code and link (owner can rotate the code)",
        "tags": [
          "squads"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rotate": {
                    "default": false,
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{32}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "inviteCode": {
                      "type": "string"
                    },
                    "deepLink": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "inviteCode",
                    "deepLink"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/squads/{id}/challenges": {
      "post": {
        "summary": "Link a pool you created to this squad so members see it in the feed and leaderboard",
        "tags": [
          "squads"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pool": {
                    "type": "string"
                  }
                },
                "required": [
                  "pool"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{32}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pool": {
                      "type": "string"
                    },
                    "squadId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "pool",
                    "squadId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/squads/{id}/leaderboard": {
      "get": {
        "summary": "Leaderboard across the squad's linked challenges",
        "tags": [
          "squads"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{32}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rows": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rank": {
                            "type": "number"
                          },
                          "wallet": {
                            "type": "string"
                          },
                          "daysCompleted": {
                            "type": "number"
                          },
                          "bestStreak": {
                            "type": "number"
                          },
                          "checkedInToday": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "rank",
                          "wallet",
                          "daysCompleted",
                          "bestStreak",
                          "checkedInToday"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "rows"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}